+++ tests/config_test.rb
@@ -16,6 +16,7 @@ class ConfigTest < Minitest::Test
     assert_equal ["test/helper.rb"], config.helper_paths
     assert_equal ["test"], config.load_paths
     assert_equal [TLDR::MOST_RECENTLY_MODIFIED_TAG], config.prepend_paths
+    assert config.warnings
   end
 
   def test_non_cli_defaults
@@ -54,15 +55,16 @@ class ConfigTest < Minitest::Test
       prepend_paths: ["a.rb:3"],
       paths: ["a.rb:3", "b.rb"],
       exclude_paths: ["c.rb:4"],
-      exclude_names: ["test_b_1"]
+      exclude_names: ["test_b_1"],
+      warnings: false
     )
 
     assert_equal <<~MSG.chomp, config.to_full_args
-      --seed 42 --verbose --reporter TLDR::Reporters::Base --helper "test_helper.rb" --load-path "app" --load-path "lib" --parallel --name "/test_*/" --name "test_it" --fail-fast --prepend "a.rb:3" --exclude-path "c.rb:4" --exclude-name "test_b_1" "a.rb:3" "b.rb"
+      --seed 42 --verbose --reporter TLDR::Reporters::Base --helper "test_helper.rb" --load-path "app" --load-path "lib" --parallel --name "/test_*/" --name "test_it" --fail-fast --prepend "a.rb:3" --exclude-path "c.rb:4" --exclude-name "test_b_1" --no-warnings "a.rb:3" "b.rb"
     MSG
 
     assert_equal <<~MSG.chomp, config.to_single_path_args("lol.rb")
-      --verbose --reporter TLDR::Reporters::Base --helper "test_helper.rb" --load-path "app" --load-path "lib" --exclude-name "test_b_1" "lol.rb"
+      --verbose --reporter TLDR::Reporters::Base --helper "test_helper.rb" --load-path "app" --load-path "lib" --exclude-name "test_b_1" --no-warnings "lol.rb"
     MSG
   end