+++ tests/fixture/warnings.rb
@@ -1,0 +1,7 @@
+class Warnful < TLDR
+  def test_warning
+  end
+
+  def test_warning # standard:disable Lint/DuplicateMethods
+  end
+end